refactor: split types to muliple files
